feat(engine-biorhythm): extract types into dedicated types.rs module (#368)
Agent-Logs-Url: https://github.com/Sheshiyer/Selemene-engine/sessions/91ef297a-8b0a-48e3-9221-4a6f1181a6d9
