--- .agent-tasks/issue-368.md
@@ -1,42 +1,0 @@
-    # Agent Task: Issue #368
-
-    > Manually dispatched fallback because Actions policy blocks PR creation.
-
-    ## Issue
-    **Title:** [V22-W1-S2-01] Extract types into dedicated types.rs module
-    **URL:** https://github.com/Sheshiyer/Selemene-engine/issues/368
-
-    ## Specification
-    <!-- plan-task-id: V22-W1-S2-01 -->
-<!-- phase: v2.2.0 -->
-
-## Extract types into dedicated types.rs module
-
-**Plan:** Specialized Engines  
-**Phase:** `v2.2.0` | **Wave:** Engine Feature Completion & Core Logic | **Swarm:** Biorhythm — Feature Gaps & Refactor  
-**Area:** `backend` | **Owner:** `Backend Eng (Rust)`  
-**Estimated Hours:** 4
-
-### Deliverable
-BiorhythmResult, CycleResult, ForecastDay, and CompatibilityResult moved to crates/engine-biorhythm/src/types.rs
-
-### Acceptance Criteria
-cargo build succeeds; all existing tests pass with types imported from types module
-
-### Validation
-types.rs exists with all public structs; lib.rs delegates via pub use
-
-### Dependencies
-- None
-
-
-    ## Agent Instructions
-    1. Read this file fully before writing any code.
-    2. Implement everything in the Deliverable section above.
-    3. Satisfy every item in the Acceptance Criteria section.
-    4. Run cargo build && cargo test and confirm all tests pass.
-    5. Delete .agent-tasks/issue-368.md when done.
-    6. Convert this PR from draft to ready-for-review.
-
-    ## Done Signal
-    Post a PR comment: agent: done -- tests passing, ready for review
