+++ crates/engine-biorhythm/src/lib.rs
@@ -4,13 +4,15 @@
 //! plus intuitive (38-day) and three composite cycles (mastery, passion, wisdom).
 //! Pure math -- no external dependencies beyond std and chrono.
 
+pub mod types;
+pub use types::*;
+
 use async_trait::async_trait;
 use chrono::{NaiveDate, Utc};
 use noesis_core::{
     CalculationMetadata, ConsciousnessEngine, EngineError, EngineInput, EngineOutput,
     ValidationResult,
 };
-use serde::{Deserialize, Serialize};
 use sha2::{Digest, Sha256};
 use std::f64::consts::PI;
 use std::time::Instant;
@@ -26,52 +28,6 @@ const INTUITIVE_PERIOD: f64 = 38.0;
 
 /// Threshold in days for declaring a zero-crossing "critical".
 const CRITICAL_THRESHOLD: f64 = 1.0;
-
-// ---------------------------------------------------------------------------
-// Result types
-// ---------------------------------------------------------------------------
-
-/// Full biorhythm calculation result.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct BiorhythmResult {
-    pub days_alive: i64,
-    pub target_date: String,
-    pub physical: CycleResult,
-    pub emotional: CycleResult,
-    pub intellectual: CycleResult,
-    pub intuitive: CycleResult,
-    pub mastery: f64,
-    pub passion: f64,
-    pub wisdom: f64,
-    pub critical_days: Vec<String>,
-    pub overall_energy: f64,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub forecast: Option<Vec<ForecastDay>>,
-}
-
-/// Result for a single biorhythm cycle.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct CycleResult {
-    pub value: f64,
-    pub percentage: f64,
-    pub phase: String,
-    pub days_until_peak: i64,
-    pub days_until_critical: i64,
-    pub is_critical: bool,
-    pub cycle_day: i64,
-}
-
-/// One day in the optional forecast window.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ForecastDay {
-    pub date: String,
-    pub days_alive: i64,
-    pub physical: f64,
-    pub emotional: f64,
-    pub intellectual: f64,
-    pub intuitive: f64,
-    pub overall_energy: f64,
-}
 
 // ---------------------------------------------------------------------------
 // Engine struct
